#CyberFraudBountyHunters
You reported a charge of $103.11 on 01/06/2025 from "Apple.com/Bill" on
your Wells Fargo account, but the receipt shows a charge of $10.80 for
"Disney" with Order ID ML4J4MNKSD, leaving $92.31 unaccounted for.

You mentioned that Apple shut down your account for not complying with
their Terms & Agreement and refused to reactivate it.

You expressed concerns about identity theft and fraud on both Apple
and Google, and plan to file a fraud report with the US Federal Trade
Commission.